+++ src/main/java/net/nabaal/majiir/realtimerender/rendering/HeightMapChunkProvider.java
@@ -9,12 +9,12 @@ public abstract class HeightMapChunkProvider extends HeightMap {
 	protected abstract void setHeightMapChunk(Coordinate chunkLocation, HeightMapChunk chunk);
 	
 	@Override
-	public final int getHeight(Coordinate point) {
+	public final byte getHeight(Coordinate point) {
 		return this.getHeightMapChunk(point.zoomOut(Coordinate.OFFSET_BLOCK_CHUNK)).getHeight(point);
 	}
 	
 	@Override
-	public final void setHeight(Coordinate block, int height) {
+	public final void setHeight(Coordinate block, byte height) {
 		Coordinate chunkLocation = block.zoomOut(Coordinate.OFFSET_BLOCK_CHUNK);
 		HeightMapChunk chunk = getHeightMapChunk(chunkLocation);
 		chunk.setHeight(block, height);
